% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OlapConnection.R
\name{OlapConnection}
\alias{OlapConnection}
\title{olapR OlapConnection Creation}
\usage{
OlapConnection(connectionString = "Data Source=localhost; Provider=MSOLAP;")
}
\arguments{
\item{connectionString}{A valid connection string for connecting to Analysis
Services}

\item{ocs}{An object of class "OlapConnection"}
}
\value{
\code{OlapConnection} returns an object of type "OlapConnection". A
warning is shown if the connection string is invalid.
}
\description{
\code{OlapConnection} constructs a "OlapConnection" object.
}
\details{
\code{OlapConnection} validates and holds an Analysis Services connection
string. By default, Analysis Services returns the first cube of the first
database. To connect to a specific database, use the Initial Catalog
parameter.
}
\examples{


# Create the connection string. For a named instance, escape the instance name: localhost\my-other-instance
cnnstr <- "Data Source=localhost; Provider=MSOLAP; initial catalog=AdventureWorksCube"
olapCnn <- OlapConnection(cnnstr)

}
\references{
For more information on Analysis Services connection strings:
\url{https://docs.microsoft.com/sql/analysis-services/instances/connection-string-properties-analysis-services}
}
\seealso{
\code{\link{Query}}, \code{\link{executeMD}},
\code{\link{execute2D}, \code{\link{explore}}}
}
